+++ config/apis/cis/v1/types.go
@@ -21,6 +21,7 @@ type VirtualServer struct {
 // VirtualServerStatus is the status of the VirtualServer resource.
 type VirtualServerStatus struct {
 	VSAddress string `json:"vsAddress,omitempty"`
+	StatusOk  string `json:"status,omitempty"`
 }
 
 // VirtualServerSpec is the spec of the VirtualServer resource.
@@ -155,6 +156,7 @@ type IngressLinkList struct {
 // +genclient
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 // +kubebuilder:validation:Optional
+// +kubebuilder:subresource:status
 
 // TransportServer defines the VirtualServer resource.
 type TransportServer struct {
@@ -162,6 +164,13 @@ type TransportServer struct {
 	metav1.ObjectMeta `json:"metadata,omitempty"`
 
 	Spec TransportServerSpec `json:"spec"`
+	Status TransportServerStatus `json:"status,omitempty"`
+}
+
+// TransportServerStatus is the status of the VirtualServer resource.
+type TransportServerStatus struct {
+	VSAddress string `json:"vsAddress,omitempty"`
+	StatusOk  string `json:"status,omitempty"`
 }
 
 // TransportServerSpec is the spec of the VirtualServer resource.